+++ langium/src/extension/main.ts
@@ -3,30 +3,30 @@ import * as vscode from 'vscode';
 import {window} from 'vscode';
 import * as path from 'node:path';
 import { LanguageClient, TransportKind } from 'vscode-languageclient/node.js';
-import { SaveImageCommand } from './image-generation/save-image-command.js';
+
 import { ImageGenerator } from './image-generation/image-generator.js';
 import { ContextManager } from './managers/context-manager.js';
 import { PreviewProvider } from './image-generation/preview-provider.js';
 import { CommandManager } from './managers/command-manager.js';
 import { EventManager, EventRunner } from './managers/event-manager.js';
+import { ConfigurationManager } from './managers/configuration-manager.js';
 let client: LanguageClient;
 
 // This function is called when the extension is activated.
 export function activate(context: vscode.ExtensionContext): void {
     client = startLanguageClient(context);
+    //create universal output channel
+    const output_channel = vscode.window.createOutputChannel("jpipe_console");
 
     //managers for updating and registration
     const command_manager = new CommandManager(context);
     const event_manager = new EventManager();
     const context_manager = new ContextManager(window.activeTextEditor);
-    
-    //create universal output channel
-    const output_channel = vscode.window.createOutputChannel("jpipe_console");
+    const configuration_manager = new ConfigurationManager(context, output_channel);
 
     //create needs for image generation
-    const save_image_command = new SaveImageCommand(context, window.activeTextEditor, output_channel);
-    const image_generator = new ImageGenerator(save_image_command, output_channel);
-    const preview_provider = new PreviewProvider(save_image_command, output_channel);
+    const image_generator = new ImageGenerator(configuration_manager, output_channel);
+    const preview_provider = new PreviewProvider(image_generator, output_channel);
 
     //register commands from classes
     command_manager.register(
@@ -36,12 +36,8 @@ export function activate(context: vscode.ExtensionContext): void {
 
     //register subscribers for events that need to monitor changes
     event_manager.register(new EventRunner(window.onDidChangeTextEditorSelection), context_manager, preview_provider);
-    event_manager.register(new EventRunner(window.onDidChangeActiveTextEditor), context_manager, save_image_command, preview_provider);
-    event_manager.register(new EventRunner(vscode.workspace.onDidChangeConfiguration), save_image_command);
-
-    vscode.workspace.onDidChangeConfiguration(() =>{
-        output_channel.appendLine("configuration changed");
-    });
+    event_manager.register(new EventRunner(window.onDidChangeActiveTextEditor), context_manager, image_generator, preview_provider);
+    event_manager.register(new EventRunner(vscode.workspace.onDidChangeConfiguration), configuration_manager);
 
     //activate listening for events
     event_manager.listen();